+++ src/userent.c
@@ -730,7 +730,7 @@ static bool laston_set(struct userrec *u, struct user_entry *e, void *buf)
 
   /* FIXME: laston sharing is disabled until a better solution is found
   if (!noshare)
-    shareout("c LASTON %s %s %li\n", u->handle, li->lastonplace ? li->lastonplace : "-", li->laston);
+    shareout("c LASTON %s %s %li\n", u->handle, li->lastonplace ? li->lastonplace : "-", (long) li->laston);
   */
 
   return 1;
